requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0

# Vision and Audio Models
timm>=0.9.0  # For DINOv2
librosa>=0.10.0
soundfile>=0.12.0

# Computer Vision
opencv-python>=4.8.0
facenet-pytorch>=2.5.0  # For MTCNN face detection
Pillow>=9.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Training and Evaluation
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
einops>=0.6.1

# Optional but Recommended
wandb>=0.15.0  # For experiment tracking
